Move storage types into own module
