+++ aquadoggo/src/db/stores/mod.rs
@@ -8,6 +8,3 @@ mod schema;
 mod task;
 #[cfg(test)]
 pub mod test_utils;
-
-pub use entry::StorageEntry;
-pub use operation::StorageOperation;
