+++ aquadoggo/src/db/types/mod.rs
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+mod entry;
+mod operation;
+
+pub use entry::StorageEntry;
+pub use operation::StorageOperation;
